% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties_sf}
\alias{counties_sf}
\title{counties_sf}
\format{
An sf data frame with 5 rows and 4 variables:
\describe{
\item{\code{country}}{Country name.}
\item{\code{state}}{State name.}
\item{\code{county}}{County name.}
\item{\code{geom}}{County geometry.}
}
}
\usage{
counties_sf
}
\description{
Sf data frame with several counties in Illinois and Indiana.
}
\keyword{datasets}
