% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_indicators.R
\name{dem_indicators}
\alias{dem_indicators}
\title{Generate Topographic Indicators from a Digital Elevation Model}
\usage{
dem_indicators(
  raster = NULL,
  dem = NULL,
  indicators = c("slope", "aspect", "TPI", "TRI", "roughness", "flowdir")
)
}
\arguments{
\item{raster}{(optional, SpatRaster) A "terra" SpatRaster object to be used as a template to crop and resample the Digital Elevation Model. If not provided, topographic indicators are computed for the complete set of tiles available in the \code{dem} argument. Default: NULL}

\item{dem}{(required, SpatRaster or path to virtual raster file) Digital Elevation Model generated by \code{\link[=dem_download]{dem_download()}} provided as a SpatRaster or as a path to a vrt file. Default: NULL}

\item{indicators}{(optional, character vector) Names of the topographic indicators to generate. Argument \code{v} of \link[terra:terrain]{terra::terrain}. These names are converted to lowercase and attached the prefix "topo_" in the output raster file. Default: c("slope", "aspect", "TPI", "TRI", "roughness", "flowdir").}
}
\value{
Multilayer SpatRaster object with the selected indicators. If \code{raster} is provided, then the output object contains all raster layers in \code{raster} as well.
}
\description{
Generates the following topographic indicators:
\itemize{
\item topo_elevation
\item topo_slope: terrain slope in degrees
\item topo_northerness: level of terrain orientation to the north.
\item topo_easterness: level of terrain orientation to the east.
\item topo_tpi: difference between the value of a cell and the mean value of its 8 surrounding cells.
\item topo_tri: Terrain Ruggedness Index is the mean of the absolute differences between the value of a cell and its 8 surrounding cells.
\item topo_roughness: difference between the maximum and the minimum value of a cell and its 8 surrounding cells.
\item topo_flowdir: direction of the greatest drop in elevation, or the smallest rise if all neighbors are higher.
}
}
\examples{

if(interactive()){

library(dem)
library(terra)

#subset one county
newton <- counties_sf[counties_sf$county == "Newton", ]

#download dem tiles for the county
dem <- dem_download(
 polygon = newton,
 resolution = 90,
 output = "raster"
)

#plot dem and polygons
plot(dem)
lines(terra::vect(newton))

#compute topographic indicators
topography <- dem_indicators(
 dem = dem
)

#plot topography
plot(topography)

}
}
