% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem_tiles}
\alias{dem_tiles}
\title{dem_tiles}
\format{
A data frame with 26475 rows and 3 variables:
\describe{
\item{\code{filename_dem_30}}{Name of the 30 m file corresponding to the tile.}
\item{\code{filename_dem_90}}{Name of the 30 m file corresponding to the tile.}
\item{\code{geometry}}{Tile geometry.}
}
}
\usage{
dem_tiles
}
\description{
Tiles of the Copernicus Digital Elevation Model
}
\details{
Modified from: \url{https://spacedata.copernicus.eu/collections/copernicus-digital-elevation-model}. Direct link to the original file: \url{https://spacedata.copernicus.eu/documents/20123/121239/GEO1988-CopernicusDEM-RP-002_GridFile_I6.0.shp.zip}
}
\keyword{datasets}
