% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jacetania}
\alias{jacetania}
\title{jacetania}
\format{
An sf data frame with 1 rows and 14 variables:
\describe{
\item{\code{objectid}}{ID object.}
\item{\code{c_comarca}}{Region code.}
\item{\code{d_comarca}}{Region name.}
\item{\code{sum_sup_of}}{Region surface in km2.}
\item{\code{fonetica}}{Phonetics of the name.}
\item{\code{ccomarca}}{Region code.}
\item{\code{scale}}{Scale.}
\item{\code{dcomarca1}}{Region name.}
\item{\code{uri_aod}}{URL data.}
\item{\code{on_date}}{Beggining date.}
\item{\code{off_date}}{Ending date.}
\item{\code{fecha_ini}}{Beggining date.}
\item{\code{fecha_fin}}{Ending date.}
\item{\code{geom}}{County geometry.}
}
}
\usage{
jacetania
}
\description{
Sf data frame with one region in Aragon(Spain).
}
\keyword{datasets}
